+++ packages/core/hooks/index.ts
@@ -19,5 +19,6 @@ export * from './useStateRef'
 export * from './useStorage'
 export * from './useToggle'
 export * from './useUgoira'
+export * from './useUnmount'
 export * from './useUser'
 export * from './useUserTags'
