+++ packages/core/hooks/useUnmount.ts
@@ -1,0 +1,14 @@
+import { useRef, useLayoutEffect } from 'react'
+
+export function useUnmount() {
+  const unmounted = useRef(true)
+
+  useLayoutEffect(() => {
+    unmounted.current = false
+    return () => {
+      unmounted.current = true
+    }
+  }, [unmounted])
+
+  return unmounted
+}
